refactor(logout): SCRUM-50 align folders and namespaces by docs/modular-monolith.md
