+++ tests/IAM/Infrastructure/Persistence/MySQL/Logout/DeleteRevokedExpiredTest.php
@@ -13,7 +13,7 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Twitter\IAM\Infrastructure\Auth\RefreshTokenHasher;
-use Twitter\IAM\Infrastructure\Persistence\MySQL\Logout\MySQLRefreshToken;
+use Twitter\IAM\Infrastructure\Persistence\MySQL\Repository\MySQLRefreshToken;
 
 #[Group('unit')]
 #[CoversClass(MySQLRefreshToken::class)]
